extract MviProfileAddress from MviProfile
