+++ lib/mvi/models/mvi_profile.rb
@@ -4,18 +4,6 @@ require 'common/models/attribute_types/date_time_string'
 
 module MVI
   module Models
-    # A vets primary address in MVI
-    class MviProfileAddress
-      include Virtus.model
-
-      attribute :street, String
-      attribute :city, String
-      attribute :state, String
-      attribute :postal_code, String
-      attribute :country, String
-    end
-
-    # A vets attributes in MVI
     class MviProfile
       include Virtus.model
