+++ lib/mvi/models/mvi_profile_address.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+require 'common/models/attribute_types/date_time_string'
+
+module MVI
+  module Models
+    class MviProfileAddress
+      include Virtus.model
+
+      attribute :street, String
+      attribute :city, String
+      attribute :state, String
+      attribute :postal_code, String
+      attribute :country, String
+    end
+  end
+end
